requirement.txt
fastapi
uvicorn
pytesseract
opencv-python-headless
pillow
numpy
